Move ServerAddress to enigma-server
