+++ enigma-server/src/main/java/cuchaz/enigma/network/ServerAddress.java
@@ -1,4 +1,4 @@
-package cuchaz.enigma.utils;
+package cuchaz.enigma.network;
 
 import java.util.Objects;
